Completed run phase refactor implementation and updated tests
