+++ src/battle.ts
@@ -94,6 +94,7 @@ export default class Battle {
   /** If the current battle is a Mystery Encounter, this will always be defined */
   public mysteryEncounter?: MysteryEncounter;
 
+  /**Tracks whether the last run attempt in battle failed*/
   public failedRunAway = false;
 
   private rngCounter = 0;
