+++ src/utils/run-utils.ts
@@ -1,0 +1,71 @@
+import type { PlayerPokemon, EnemyPokemon } from "#app/field/pokemon";
+import type { NumberHolder } from "#app/utils/common";
+import type Pokemon from "#app/field/pokemon";
+import { Stat } from "#enums/stat";
+
+/**
+ * Calculates the chance for the player's team to successfully run away from battle.
+ *
+ * @param playerField       The player's currently active Pokémon
+ * @param enemyField        The enemy team's Pokémon
+ * @param escapeChance      A mutable NumberHolder to be assigned the final escape chance value
+ * @param escapeAttempts    The number of previous escape attempts made in the battle
+ */
+export function calculateEscapeChance(
+  playerField: PlayerPokemon[],
+  enemyField: EnemyPokemon[],
+  escapeChance: NumberHolder,
+  escapeAttempts: number,
+) {
+  /** Sum of the speed of all enemy pokemon on the field */
+  const enemySpeed = enemyField.reduce(
+    (total: number, enemyPokemon: Pokemon) => total + enemyPokemon.getStat(Stat.SPD),
+    0,
+  );
+  /** Sum of the speed of all player pokemon on the field */
+  const playerSpeed = playerField.reduce(
+    (total: number, playerPokemon: Pokemon) => total + playerPokemon.getStat(Stat.SPD),
+    0,
+  );
+
+  /*  The way the escape chance works is by looking at the difference between your speed and the enemy field's average speed as a ratio. The higher this ratio, the higher your chance of success.
+   *  However, there is a cap for the ratio of your speed vs enemy speed which beyond that point, you won't gain any advantage. It also looks at how many times you've tried to escape.
+   *  Again, the more times you've tried to escape, the higher your odds of escaping. Bosses and non-bosses are calculated differently - bosses are harder to escape from vs non-bosses
+   *  Finally, there's a minimum and maximum escape chance as well so that escapes aren't guaranteed, yet they are never 0 either.
+   *  The percentage chance to escape from a pokemon for both bosses and non bosses is linear and based on the minimum and maximum chances, and the speed ratio cap.
+   *
+   *  At the time of writing, these conditions should be met:
+   *   - The minimum escape chance should be 5% for bosses and non bosses
+   *   - Bosses should have a maximum escape chance of 25%, whereas non-bosses should be 95%
+   *   - The bonus per previous escape attempt should be 2% for bosses and 10% for non-bosses
+   *   - The speed ratio cap should be 6x for bosses and 4x for non-bosses
+   *   - The "default" escape chance when your speed equals the enemy speed should be 8.33% for bosses and 27.5% for non-bosses
+   *
+   *  From the above, we can calculate the below values
+   */
+
+  let isBoss = false;
+  for (let e = 0; e < enemyField.length; e++) {
+    isBoss = isBoss || enemyField[e].isBoss(); // this line checks if any of the enemy pokemon on the field are bosses; if so, the calculation for escaping is different
+  }
+
+  /** The ratio between the speed of your active pokemon and the speed of the enemy field */
+  const speedRatio = playerSpeed / enemySpeed;
+  /** The max ratio before escape chance stops increasing. Increased if there is a boss on the field */
+  const speedCap = isBoss ? 6 : 4;
+  /** Minimum percent chance to escape */
+  const minChance = 5;
+  /** Maximum percent chance to escape. Decreased if a boss is on the field */
+  const maxChance = isBoss ? 25 : 95;
+  /** How much each escape attempt increases the chance of the next attempt. Decreased if a boss is on the field */
+  const escapeBonus = isBoss ? 2 : 10;
+  /** Slope of the escape chance curve */
+  const escapeSlope = (maxChance - minChance) / speedCap;
+
+  // This will calculate the escape chance given all of the above and clamp it to the range of [`minChance`, `maxChance`]
+  escapeChance.value = Phaser.Math.Clamp(
+    Math.round(escapeSlope * speedRatio + minChance + escapeBonus * escapeAttempts),
+    minChance,
+    maxChance,
+  );
+}
